refactor: reorganize MCP entities and controllers into a dedicated mcp-server directory
