+++ packages/server/api/src/app/database/database-connection.ts
@@ -42,7 +42,7 @@ import { FlowVersionEntity } from '../flows/flow-version/flow-version-entity'
 import { FolderEntity } from '../flows/folder/folder.entity'
 import { IssueEntity } from '../flows/issues/issues-entity'
 import { DatabaseType, system } from '../helper/system/system'
-import { McpEntity } from '../mcp/mcp-entity'
+import { McpEntity } from '../mcp/mcp-server/mcp-entity'
 import { McpRunEntity } from '../mcp/mcp-run/mcp-run.entity'
 import { McpToolEntity } from '../mcp/tool/mcp-tool.entity'
 import { PieceMetadataEntity } from '../pieces/piece-metadata-entity'
