+++ packages/server/api/src/app/mcp/mcp-module.ts
@@ -2,7 +2,7 @@ import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox'
 import { mcpRunController } from './mcp-run/mcp-run.controller'
 import { mcpSessionManager } from './mcp-server/mcp-session-manager'
 import { mcpSseController } from './mcp-server/mcp-sse-controller'
-import { mcpServerController } from './mcp-server-controller'
+import { mcpServerController } from './mcp-server/mcp-server-controller'
 
 export const mcpModule: FastifyPluginAsyncTypebox = async (app) => {
     await app.register(mcpServerController, { prefix: '/v1/mcp-servers' })
