+++ packages/server/api/src/app/mcp/mcp-server/mcp-entity.ts
@@ -1,6 +1,6 @@
 import { Agent, McpWithTools } from '@activepieces/shared'
 import { EntitySchema } from 'typeorm'
-import { ApIdSchema, BaseColumnSchemaPart } from '../database/database-common'
+import { ApIdSchema, BaseColumnSchemaPart } from '../../database/database-common'
 
 type McpWithToolsWithSchema = McpWithTools & {  
     agent: Agent
