+++ packages/server/api/src/app/mcp/mcp-server/mcp-server-controller.ts
@@ -1,9 +1,9 @@
 import { apId, ApId, CreateMcpRequestBody, ListMcpsRequest, McpWithTools, Nullable, Permission, PlatformUsageMetric, PrincipalType, SeekPage, SERVICE_KEY_SECURITY_OPENAPI, UpdateMcpRequestBody } from '@activepieces/shared'
 import { FastifyPluginAsyncTypebox, Type } from '@fastify/type-provider-typebox'
 import { StatusCodes } from 'http-status-codes'
-import { entitiesMustBeOwnedByCurrentProject } from '../authentication/authorization'
-import { PlatformPlanHelper } from '../ee/platform/platform-plan/platform-plan-helper'
-import { mcpService } from './mcp-service'
+import { entitiesMustBeOwnedByCurrentProject } from '../../authentication/authorization'
+import { PlatformPlanHelper } from '../../ee/platform/platform-plan/platform-plan-helper'
+import { mcpService } from '../mcp-service'
 
 const DEFAULT_PAGE_SIZE = 10
